feat(apps-api-codelab): use typescript json schema lib
